+++ tsdb/engine.go
@@ -33,6 +33,8 @@ type Engine interface {
 	DeleteSeries(keys []string) error
 	DeleteMeasurement(name string, seriesKeys []string) error
 	SeriesCount() (n int, err error)
+
+	io.WriterTo
 }
 
 // NewEngineFunc creates a new engine.